// core/models/TipoDeArticulo.ts
export interface TipoDeArticulo{
    otrosF1: string;
    otrosN1: number;
    otrosN2: number;
    otrosC1: null;
    otrosC2: null;
    otrosC3: null;
    otrosC4: null;
    otrosT1: null;
    usuario: number;
    equipo: string;
    id: number;
    nombre: string;
    suspendido: boolean;
    manejaInventario: boolean;
}